@@ -1,17 +1,19 @@
 {
 	"name": "@szmarczak/http-timer",
-	"version": "1.1.2",
+	"version": "4.0.5",
 	"description": "Timings for HTTP requests",
-	"main": "source",
+	"main": "dist/source",
 	"engines": {
-		"node": ">=6"
+		"node": ">=10"
 	},
 	"scripts": {
-		"test": "xo && nyc ava",
+		"test": "xo && tsc --noEmit && nyc ava",
+		"build": "del-cli dist && tsc",
+		"prepare": "npm run build",
 		"coveralls": "nyc report --reporter=text-lcov | coveralls"
 	},
 	"files": [
-		"source"
+		"dist/source"
 	],
 	"keywords": [
 		"http",
@@ -29,19 +31,44 @@
 		"url": "https://github.com/szmarczak/http-timer/issues"
 	},
 	"homepage": "https://github.com/szmarczak/http-timer#readme",
-	"xo": {
-		"rules": {
-			"unicorn/filename-case": "camelCase"
-		}
+	"dependencies": {
+		"defer-to-connect": "^2.0.0"
 	},
 	"devDependencies": {
-		"ava": "^0.25.0",
-		"coveralls": "^3.0.2",
-		"p-event": "^2.1.0",
-		"nyc": "^12.0.2",
-		"xo": "^0.22.0"
+		"@ava/typescript": "^1.1.0",
+		"@sindresorhus/tsconfig": "^0.7.0",
+		"@types/node": "^13.5.1",
+		"@typescript-eslint/eslint-plugin": "^2.18.0",
+		"@typescript-eslint/parser": "^2.18.0",
+		"ava": "^3.2.0",
+		"coveralls": "^3.0.9",
+		"del-cli": "^3.0.0",
+		"eslint-config-xo-typescript": "^0.24.1",
+		"nyc": "^15.0.0",
+		"p-event": "^4.1.0",
+		"typescript": "^3.7.5",
+		"xo": "^0.25.3"
 	},
-	"dependencies": {
-		"defer-to-connect": "^1.0.1"
+	"types": "dist/source",
+	"xo": {
+		"extends": "xo-typescript",
+		"extensions": [
+			"ts"
+		]
+	},
+	"nyc": {
+		"extension": [
+			".ts"
+		],
+		"exclude": [
+			"**/tests/**"
+		]
+	},
+	"ava": {
+		"typescript": {
+			"rewritePaths": {
+				"tests/": "dist/tests/"
+			}
+		}
 	}
 }
