@@ -1,6 +1,6 @@
 {
   "name": "babel-preset-gatsby",
-  "version": "0.2.35",
+  "version": "0.4.7",
   "author": "Philipp Spiess <hello@philippspiess.com>",
   "repository": {
     "type": "git",
@@ -9,19 +9,19 @@
   },
   "homepage": "https://github.com/gatsbyjs/gatsby/tree/master/packages/babel-preset-gatsby#readme",
   "dependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.7.4",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.7.4",
-    "@babel/plugin-proposal-optional-chaining": "^7.7.5",
-    "@babel/plugin-syntax-dynamic-import": "^7.7.4",
-    "@babel/plugin-transform-runtime": "^7.7.6",
-    "@babel/plugin-transform-spread": "^7.7.4",
-    "@babel/preset-env": "^7.7.6",
-    "@babel/preset-react": "^7.7.4",
-    "@babel/runtime": "^7.7.6",
-    "babel-plugin-dynamic-import-node": "^2.3.0",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.8.3",
+    "@babel/plugin-proposal-optional-chaining": "^7.9.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.9.6",
+    "@babel/plugin-transform-spread": "^7.8.3",
+    "@babel/preset-env": "^7.9.6",
+    "@babel/preset-react": "^7.9.4",
+    "@babel/runtime": "^7.9.6",
+    "babel-plugin-dynamic-import-node": "^2.3.3",
     "babel-plugin-macros": "^2.8.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
-    "gatsby-core-utils": "^1.0.33"
+    "gatsby-core-utils": "^1.3.3"
   },
   "peerDependencies": {
     "@babel/core": "^7.0.0"
@@ -29,18 +29,18 @@
   "license": "MIT",
   "main": "index.js",
   "scripts": {
-    "build": "babel src --out-dir . --ignore **/__tests__ --ignore **/utils/path-serializer.ts --extensions \".ts,.js\"",
+    "build": "babel src --out-dir . --ignore \"**/__tests__\" --ignore \"**/utils/path-serializer.ts\" --extensions \".ts,.js\"",
     "prepare": "cross-env NODE_ENV=production npm run build",
-    "watch": "babel -w src --out-dir . --ignore **/__tests__ --ignore **/utils/path-serializer.ts --extensions \".ts,.js\""
+    "watch": "babel -w src --out-dir . --ignore \"**/__tests__\" --ignore \"**/utils/path-serializer.ts\" --extensions \".ts,.js\""
   },
   "devDependencies": {
-    "@babel/cli": "^7.7.5",
-    "babel-preset-gatsby-package": "^0.2.17",
+    "@babel/cli": "^7.8.4",
+    "babel-preset-gatsby-package": "^0.4.2",
     "cross-env": "^5.2.1",
     "slash": "^3.0.0"
   },
   "engines": {
-    "node": ">=8.0.0"
+    "node": ">=10.13.0"
   },
-  "gitHead": "edb1423009c6a5cfc0e811331abde8716f07f306"
+  "gitHead": "e3c4efcbae6f5eab377a40bc88f45511f7c02c80"
 }
