@@ -1,53 +1,51 @@
 {
-  "name": "slice-ansi",
-  "version": "1.0.0",
-  "description": "Slice a string with ANSI escape codes",
-  "license": "MIT",
-  "repository": "chalk/slice-ansi",
-  "author": {
-    "name": "David Caccavella",
-    "email": "threedeecee@gmail.com"
-  },
-  "engines": {
-    "node": ">=4"
-  },
-  "scripts": {
-    "test": "xo && ava"
-  },
-  "files": [
-    "index.js"
-  ],
-  "keywords": [
-    "slice",
-    "string",
-    "ansi",
-    "styles",
-    "color",
-    "colour",
-    "colors",
-    "terminal",
-    "console",
-    "cli",
-    "tty",
-    "escape",
-    "formatting",
-    "rgb",
-    "256",
-    "shell",
-    "xterm",
-    "log",
-    "logging",
-    "command-line",
-    "text"
-  ],
-  "dependencies": {
-    "is-fullwidth-code-point": "^2.0.0"
-  },
-  "devDependencies": {
-    "ava": "*",
-    "chalk": "^2.0.1",
-    "random-item": "^1.0.0",
-    "strip-ansi": "^4.0.0",
-    "xo": "*"
-  }
+	"name": "slice-ansi",
+	"version": "2.1.0",
+	"description": "Slice a string with ANSI escape codes",
+	"license": "MIT",
+	"repository": "chalk/slice-ansi",
+	"engines": {
+		"node": ">=6"
+	},
+	"scripts": {
+		"test": "xo && ava"
+	},
+	"files": [
+		"index.js"
+	],
+	"keywords": [
+		"slice",
+		"string",
+		"ansi",
+		"styles",
+		"color",
+		"colour",
+		"colors",
+		"terminal",
+		"console",
+		"cli",
+		"tty",
+		"escape",
+		"formatting",
+		"rgb",
+		"256",
+		"shell",
+		"xterm",
+		"log",
+		"logging",
+		"command-line",
+		"text"
+	],
+	"dependencies": {
+		"ansi-styles": "^3.2.0",
+		"astral-regex": "^1.0.0",
+		"is-fullwidth-code-point": "^2.0.0"
+	},
+	"devDependencies": {
+		"ava": "^1.1.0",
+		"chalk": "^2.4.2",
+		"random-item": "^1.0.0",
+		"strip-ansi": "^5.0.0",
+		"xo": "^0.24.0"
+	}
 }
