@@ -1,6 +1,6 @@
 {
 	"name": "to-readable-stream",
-	"version": "1.0.0",
+	"version": "2.1.0",
 	"description": "Convert a string/Buffer/Uint8Array to a readable stream",
 	"license": "MIT",
 	"repository": "sindresorhus/to-readable-stream",
@@ -10,13 +10,14 @@
 		"url": "sindresorhus.com"
 	},
 	"engines": {
-		"node": ">=6"
+		"node": ">=8"
 	},
 	"scripts": {
-		"test": "xo && ava"
+		"test": "xo && ava && tsd"
 	},
 	"files": [
-		"index.js"
+		"index.js",
+		"index.d.ts"
 	],
 	"keywords": [
 		"stream",
@@ -33,8 +34,9 @@
 		"pull"
 	],
 	"devDependencies": {
-		"ava": "*",
-		"get-stream": "^3.0.0",
-		"xo": "*"
+		"ava": "^1.4.1",
+		"get-stream": "^5.0.0",
+		"tsd": "^0.7.2",
+		"xo": "^0.24.0"
 	}
 }
