@@ -2,6 +2,11 @@
 const preferDefault = m => m && m.default || m
 
 exports.components = {
+  "component---src-templates-codepen-example-js": () => import("./../src/templates/codepen-example.js" /* webpackChunkName: "component---src-templates-codepen-example-js" */),
+  "component---src-templates-docs-js": () => import("./../src/templates/docs.js" /* webpackChunkName: "component---src-templates-docs-js" */),
+  "component---src-templates-community-js": () => import("./../src/templates/community.js" /* webpackChunkName: "component---src-templates-community-js" */),
+  "component---src-templates-blog-js": () => import("./../src/templates/blog.js" /* webpackChunkName: "component---src-templates-blog-js" */),
+  "component---src-templates-tutorial-js": () => import("./../src/templates/tutorial.js" /* webpackChunkName: "component---src-templates-tutorial-js" */),
   "component---src-pages-404-js": () => import("./../src/pages/404.js" /* webpackChunkName: "component---src-pages-404-js" */),
   "component---src-pages-acknowledgements-html-js": () => import("./../src/pages/acknowledgements.html.js" /* webpackChunkName: "component---src-pages-acknowledgements-html-js" */),
   "component---src-pages-blog-all-html-js": () => import("./../src/pages/blog/all.html.js" /* webpackChunkName: "component---src-pages-blog-all-html-js" */),
@@ -9,11 +14,6 @@ exports.components = {
   "component---src-pages-index-js": () => import("./../src/pages/index.js" /* webpackChunkName: "component---src-pages-index-js" */),
   "component---src-pages-jsx-compiler-html-js": () => import("./../src/pages/jsx-compiler.html.js" /* webpackChunkName: "component---src-pages-jsx-compiler-html-js" */),
   "component---src-pages-languages-js": () => import("./../src/pages/languages.js" /* webpackChunkName: "component---src-pages-languages-js" */),
-  "component---src-pages-versions-js": () => import("./../src/pages/versions.js" /* webpackChunkName: "component---src-pages-versions-js" */),
-  "component---src-templates-blog-js": () => import("./../src/templates/blog.js" /* webpackChunkName: "component---src-templates-blog-js" */),
-  "component---src-templates-codepen-example-js": () => import("./../src/templates/codepen-example.js" /* webpackChunkName: "component---src-templates-codepen-example-js" */),
-  "component---src-templates-community-js": () => import("./../src/templates/community.js" /* webpackChunkName: "component---src-templates-community-js" */),
-  "component---src-templates-docs-js": () => import("./../src/templates/docs.js" /* webpackChunkName: "component---src-templates-docs-js" */),
-  "component---src-templates-tutorial-js": () => import("./../src/templates/tutorial.js" /* webpackChunkName: "component---src-templates-tutorial-js" */)
+  "component---src-pages-versions-js": () => import("./../src/pages/versions.js" /* webpackChunkName: "component---src-pages-versions-js" */)
 }
 
