@@ -8,8 +8,6 @@ var _reactDom = _interopRequireDefault(require("react-dom"));
 
 var _domready = _interopRequireDefault(require("@mikaelkristiansson/domready"));
 
-var _socket = _interopRequireDefault(require("socket.io-client"));
-
 var _socketIo = _interopRequireDefault(require("./socketIo"));
 
 var _emitter = _interopRequireDefault(require("./emitter"));
@@ -40,23 +38,6 @@ window.___loader = _loader.publicLoader; // Let the site/plugins run code very e
       window.location.reload();
     });
   }
-
-  fetch(`/___services`).then(res => res.json()).then(services => {
-    if (services.developstatusserver) {
-      const parentSocket = (0, _socket.default)(`${window.location.protocol}//${window.location.hostname}:${services.developstatusserver.port}`);
-      parentSocket.on(`develop:needs-restart`, msg => {
-        if (window.confirm(`The develop process needs to be restarted for the changes to ${msg.dirtyFile} to be applied.\nDo you want to restart the develop process now?`)) {
-          parentSocket.once(`develop:is-starting`, msg => {
-            window.location.reload();
-          });
-          parentSocket.once(`develop:started`, msg => {
-            window.location.reload();
-          });
-          parentSocket.emit(`develop:restart`);
-        }
-      });
-    }
-  });
   /**
    * Service Workers are persistent by nature. They stick around,
    * serving a cached version of the site if they aren't removed.
@@ -65,6 +46,7 @@ window.___loader = _loader.publicLoader; // Let the site/plugins run code very e
    *
    * Let's warn if we find service workers in development.
    */
+
 
   if (`serviceWorker` in navigator) {
     navigator.serviceWorker.getRegistrations().then(registrations => {
@@ -79,7 +61,7 @@ window.___loader = _loader.publicLoader; // Let the site/plugins run code very e
 
     let Root = preferDefault(require(`./root`));
     (0, _domready.default)(() => {
-      renderer( /*#__PURE__*/_react.default.createElement(Root, null), rootElement, () => {
+      renderer(_react.default.createElement(Root, null), rootElement, () => {
         (0, _apiRunnerBrowser.apiRunner)(`onInitialClientRender`);
       });
     });
