@@ -10,15 +10,15 @@ var _react = _interopRequireDefault(require("react"));
 var _propTypes = _interopRequireDefault(require("prop-types"));
 
 function HTML(props) {
-  return /*#__PURE__*/_react.default.createElement("html", props.htmlAttributes, /*#__PURE__*/_react.default.createElement("head", null, /*#__PURE__*/_react.default.createElement("meta", {
+  return _react.default.createElement("html", props.htmlAttributes, _react.default.createElement("head", null, _react.default.createElement("meta", {
     charSet: "utf-8"
-  }), /*#__PURE__*/_react.default.createElement("meta", {
+  }), _react.default.createElement("meta", {
     httpEquiv: "x-ua-compatible",
     content: "ie=edge"
-  }), /*#__PURE__*/_react.default.createElement("meta", {
+  }), _react.default.createElement("meta", {
     name: "viewport",
     content: "width=device-width, initial-scale=1, shrink-to-fit=no"
-  }), props.headComponents), /*#__PURE__*/_react.default.createElement("body", props.bodyAttributes, props.preBodyComponents, /*#__PURE__*/_react.default.createElement("div", {
+  }), props.headComponents), _react.default.createElement("body", props.bodyAttributes, props.preBodyComponents, _react.default.createElement("div", {
     key: `body`,
     id: "___gatsby",
     dangerouslySetInnerHTML: {
