@@ -21,8 +21,7 @@ class EnsureResources extends _react.default.Component {
       pageResources
     } = props;
     this.state = {
-      location: { ...location
-      },
+      location: Object.assign({}, location),
       pageResources: pageResources || _loader.default.loadPageSync(location.pathname)
     };
   }
@@ -35,14 +34,12 @@ class EnsureResources extends _react.default.Component {
 
       return {
         pageResources,
-        location: { ...location
-        }
+        location: Object.assign({}, location)
       };
     }
 
     return {
-      location: { ...location
-      }
+      location: Object.assign({}, location)
     };
   }
 
@@ -50,8 +47,7 @@ class EnsureResources extends _react.default.Component {
     _loader.default.loadPage(rawPath).then(pageResources => {
       if (pageResources && pageResources.status !== _loader.PageResourceStatus.Error) {
         this.setState({
-          location: { ...window.location
-          },
+          location: Object.assign({}, window.location),
           pageResources
         });
       } else {
