@@ -43,7 +43,7 @@ function StaticQueryDataRenderer({
   render
 }) {
   const finalData = data ? data.data : staticQueryData[query] && staticQueryData[query].data;
-  return /*#__PURE__*/_react.default.createElement(_react.default.Fragment, null, finalData && render(finalData), !finalData && /*#__PURE__*/_react.default.createElement("div", null, "Loading (StaticQuery)"));
+  return _react.default.createElement(_react.default.Fragment, null, finalData && render(finalData), !finalData && _react.default.createElement("div", null, "Loading (StaticQuery)"));
 }
 
 const StaticQuery = props => {
@@ -53,7 +53,7 @@ const StaticQuery = props => {
     render,
     children
   } = props;
-  return /*#__PURE__*/_react.default.createElement(StaticQueryContext.Consumer, null, staticQueryData => /*#__PURE__*/_react.default.createElement(StaticQueryDataRenderer, {
+  return _react.default.createElement(StaticQueryContext.Consumer, null, staticQueryData => _react.default.createElement(StaticQueryDataRenderer, {
     data: data,
     query: query,
     render: render || children,
@@ -68,19 +68,7 @@ const useStaticQuery = query => {
     throw new Error(`You're likely using a version of React that doesn't support Hooks\n` + `Please update React and ReactDOM to 16.8.0 or later to use the useStaticQuery hook.`);
   }
 
-  const context = _react.default.useContext(StaticQueryContext); // query is a stringified number like `3303882` when wrapped with graphql, If a user forgets
-  // to wrap the query in a grqphql, then casting it to a Number results in `NaN` allowing us to
-  // catch the misuse of the API and give proper direction
-
-
-  if (isNaN(Number(query))) {
-    throw new Error(`useStaticQuery was called with a string but expects to be called using \`graphql\`. Try this:
-
-import { useStaticQuery, graphql } from 'gatsby';
-
-useStaticQuery(graphql\`${query}\`);
-`);
-  }
+  const context = _react.default.useContext(StaticQueryContext);
 
   if (context[query] && context[query].data) {
     return context[query].data;
