@@ -15,19 +15,18 @@ var _loader = require("./loader");
 
 var _apiRunnerBrowser = require("./api-runner-browser");
 
-// Renders page
 class PageRenderer extends _react.default.Component {
   render() {
-    const props = { ...this.props,
+    const props = Object.assign({}, this.props, {
       pathContext: this.props.pageContext
-    };
+    });
     const [replacementElement] = (0, _apiRunnerBrowser.apiRunner)(`replaceComponentRenderer`, {
       props: this.props,
       loader: _loader.publicLoader
     });
-    const pageElement = replacementElement || (0, _react.createElement)(this.props.pageResources.component, { ...props,
+    const pageElement = replacementElement || (0, _react.createElement)(this.props.pageResources.component, Object.assign({}, props, {
       key: this.props.path || this.props.pageResources.page.path
-    });
+    }));
     const wrappedPage = (0, _apiRunnerBrowser.apiRunner)(`wrapPageElement`, {
       element: pageElement,
       props
