@@ -1,11 +1,11 @@
 "use strict";
 
 exports.__esModule = true;
-exports.testRequireError = void 0;
+exports.default = void 0;
 
 // This module is also copied into the .cache directory some modules copied there
 // from cache-dir can also use this module.
-const testRequireError = (moduleName, err) => {
+var _default = (moduleName, err) => {
   // PnP will return the following code when a require is allowed per the
   // dependency tree rules but the requested file doesn't exist
   if (err.code === `QUALIFIED_PATH_RESOLUTION_FAILED` || err.pnpCode === `QUALIFIED_PATH_RESOLUTION_FAILED`) {
@@ -17,5 +17,5 @@ const testRequireError = (moduleName, err) => {
   return regex.test(firstLine);
 };
 
-exports.testRequireError = testRequireError;
+exports.default = _default;
 //# sourceMappingURL=test-require-error.js.map
